+++ x-pack/plugins/enterprise_search/public/applications/index.tsx
@@ -6,15 +6,13 @@
 
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { BrowserRouter, Route, Redirect } from 'react-router-dom';
+import { Router, Route, Redirect } from 'react-router-dom';
 
 import { CoreStart, AppMountParams, HttpHandler } from 'src/core/public';
 import { ClientConfigType, PluginsSetup } from '../plugin';
 import { TSetBreadcrumbs } from './shared/kibana_breadcrumbs';
 import { ILicense } from '../../../../licensing/public';
 import { LicenseProvider } from './shared/licensing';
-
-import { AppSearch } from './app_search';
 
 export interface IKibanaContext {
   enterpriseSearchUrl?: string;
@@ -25,7 +23,14 @@ export interface IKibanaContext {
 
 export const KibanaContext = React.createContext();
 
+/**
+ * This file serves as a reusable wrapper to share Kibana-level context and other helpers
+ * between various Enterprise Search plugins (e.g. AppSearch, WorkplaceSearch, ES landing page)
+ * which should be imported and passed in as the first param in plugin.ts.
+ */
+
 export const renderApp = (
+  App: React.Element,
   core: CoreStart,
   params: AppMountParams,
   config: ClientConfigType,
@@ -41,16 +46,9 @@ export const renderApp = (
       }}
     >
       <LicenseProvider>
-        <BrowserRouter basename={params.appBasePath}>
-          <Route exact path="/">
-            {/* This will eventually contain an Enterprise Search landing page,
-            and we'll also actually have a /workplace_search route */}
-            <Redirect to="/app_search" />
-          </Route>
-          <Route path="/app_search">
-            <AppSearch />
-          </Route>
-        </BrowserRouter>
+        <Router history={params.history}>
+          <App />
+        </Router>
       </LicenseProvider>
     </KibanaContext.Provider>,
     params.element
